requirements-lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
bandit==1.9.3
black==26.1.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorlog==6.10.1
confection==0.1.5
coverage==7.13.2
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cymem==2.0.13
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dparse==0.6.4
execnet==2.1.2
factory_boy==3.3.3
faiss-gpu-cu12==1.13.2
Faker==40.1.2
fastuuid==0.14.0
filelock==3.20.3
flake8==7.3.0
freezegun==1.5.5
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.7
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.8
langchain-text-splitters==1.1.0
langdetect==1.0.9
langsmith==0.6.8
librt==0.7.8
litellm==1.81.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.1
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
nodeenv==1.10.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.16.0
orjson==3.11.7
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
post_slug==1.0.1
pre_commit==4.5.1
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-mock==3.15.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.25.8
rich==14.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
safety==3.7.0
safety-schemas==0.0.16
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.2.0
starlette==0.52.1
stevedore==5.6.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
virtualenv==20.36.1
wasabi==1.1.3
weasel==0.4.3
wheel==0.46.3
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
